requirements.txt
alembic==1.12.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==2.2.5
Flask-Migrate==4.0.5
flask-sqlalchemy==3.0.5
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pandas==1.3.5
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
werkzeug==2.2.3
zipp==3.15.0
